Deleted AgeCalculator - Not needed anymore Annual tax tests completed Annual tax calculator completed and tested Mersenne prime calculator completed and tested Solution completed and ready for acceptance testing by Fedgroup

+++ FGEvaluation/AnnualTaxCalculator/Models/AgeBasedRelief.cs
@@ -3,7 +3,7 @@
     public class AgeBasedRelief
     {
         public int LowerAgeLimit { get; set; }
-        public double MonthlySalaryRelief { get; set; }
+        public decimal MonthlySalaryRelief { get; set; }
         public decimal PercentageOfDeduction { get; set; }
     }
 }
